Fix text filter properties and rename files
